feat: enhanced version of the history
