+++ src/components/screens/HistoryScreen/index.ts
@@ -1,2 +1,9 @@
-export { default } from "components/screens/HistoryScreen/HistoryScreen";
-export * from "components/screens/HistoryScreen/types";
+// Export types
+export * from "./types";
+
+// Export components
+export { default as HistoryItem } from "./HistoryItem";
+export { default as HistoryScreen } from "./HistoryScreen";
+
+// Export mappers
+export { mapHistoryItemData } from "./mappers";
