+++ src/components/screens/HistoryScreen/types.ts
@@ -1,3 +1,4 @@
+/* eslint-disable @typescript-eslint/no-explicit-any */
 import { Horizon } from "@stellar/stellar-sdk";
 import { SorobanTokenInterface } from "helpers/soroban";
 
@@ -17,6 +18,55 @@ export enum TransactionStatus {
   FAILED = "failed",
 }
 
+// Base interfaces for transaction details
+export interface CreateAccountDetailsType {
+  isCreatingExternalAccount: boolean;
+  accountPublicKey: string;
+  startingBalance: string;
+}
+
+export interface SwapDetailsType {
+  sourceAssetCode: string;
+  sourceAssetIssuer: string;
+  destinationAssetCode: string;
+  destinationAssetIssuer: string;
+  sourceAssetType: string;
+  destinationAssetType: string;
+  sourceAmount: string;
+  destinationAmount: string;
+}
+
+export interface PaymentDetailsType {
+  assetCode: string;
+  assetIssuer: string;
+  assetType: string;
+  amount: string;
+  from: string;
+  to: string;
+}
+
+export interface TokenTransferDetailsType {
+  from: string;
+  to: string;
+  amount: string;
+}
+
+export interface TokenMintDetailsType {
+  from: string;
+  to: string;
+  amount: string;
+}
+
+export interface ContractDetailsType {
+  sorobanTokenInterface: SorobanTokenInterface;
+  contractAddress: string;
+  contractName?: string;
+  contractSymbol?: string;
+  contractDecimals?: number;
+  transferDetails?: TokenTransferDetailsType;
+  mintDetails?: TokenMintDetailsType;
+}
+
 export interface TransactionDetails {
   operation: Horizon.ServerApi.OperationRecord;
   transactionTitle: string;
@@ -26,44 +76,29 @@ export interface TransactionDetails {
   status: TransactionStatus;
   IconComponent: React.ReactNode;
   ActionIconComponent: React.ReactNode;
-  createAccountDetails?: {
-    isCreatingExternalAccount: boolean;
-    accountPublicKey: string;
-    startingBalance: string;
-  };
-  swapDetails?: {
-    sourceAssetCode: string;
-    sourceAssetIssuer: string;
-    destinationAssetCode: string;
-    destinationAssetIssuer: string;
-    sourceAssetType: string;
-    destinationAssetType: string;
-    sourceAmount: string;
-    destinationAmount: string;
-  };
-  paymentDetails?: {
-    assetCode: string;
-    assetIssuer: string;
-    assetType: string;
-    amount: string;
-    from: string;
-    to: string;
-  };
-  contractDetails?: {
-    sorobanTokenInterface: SorobanTokenInterface;
-    contractAddress: string;
-    contractName?: string;
-    contractSymbol?: string;
-    contractDecimals?: number;
-    transferDetails?: {
-      from: string;
-      to: string;
-      amount: string;
-    };
-    mintDetails?: {
-      from: string;
-      to: string;
-      amount: string;
-    };
-  };
+  createAccountDetails?: CreateAccountDetailsType;
+  swapDetails?: SwapDetailsType;
+  paymentDetails?: PaymentDetailsType;
+  contractDetails?: ContractDetailsType;
+}
+
+// Additional types for HistoryItem component
+export interface HistoryItemData {
+  transactionDetails: TransactionDetails;
+  rowText: string;
+  actionText: string | null;
+  ActionIconComponent: React.ReactElement | null;
+  dateText: string | null;
+  amountText: string | null;
+  IconComponent: React.ReactElement | null;
+  transactionStatus: TransactionStatus;
+  isAddingFunds: boolean | null;
+}
+
+export interface HistoryItemProps {
+  accountBalances: any; // Using any here to match existing code
+  operation: any; // Using any here to match existing code
+  publicKey: string;
+  networkDetails: any; // Using any here to match existing code
+  handleTransactionDetails: (transactionDetail: TransactionDetails) => void;
 }
